LayoutTests/http/tests/security/xssAuditor/script-tag-with-source-data-url4-expected.txt

CONSOLE MESSAGE: line 3: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-property.pl?q=%22%3E%3Cscript%20src%3ddata:,alert(1)%3bhey%%22' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
